Updates for Numpy 2.0 (Part 1) (#177)
* add np2 flag

* resolve AxisError move

* np.bool_ -> bool

* np.NaN -> np.nan

* np.Inf -> np.inf

* np.NINF -> -np.inf

* don't use private np._core APIs

* switch to public APIs

* msort was removed entirely

* don't rely on OOB integer casting

* uint0 -> uintp

* itemset was removed entirely

* use proper parametrization for ufunc tests

* give up on test_initial_list exception matching for prod

* remove msort from comparison

* remove extraneous singleton cases

* mypy

* make eager FFT error display errant dtype values

* docs updates for removed np2 things

* match prod exception more closely

* fix fft complex inputs

* don't rely on implicit OOB casting

+++ cunumeric/_module/creation_shape.py
@@ -396,6 +396,6 @@ def full_like(
     else:
         dtype = a.dtype
     result = empty_like(a, dtype=dtype, shape=shape)
-    val = np.array(value, dtype=result.dtype)
+    val = np.array(value).astype(dtype)
     result._thunk.fill(val)
     return result
